+++ releasenotes.md
@@ -9,7 +9,8 @@
 - `"build-dir"` option now available for `project.json`, added to project. #2323
 - Allow `..` ranges to use "a..a-1" in order to express zero length.
 - Disallow aliasing of `@local` symbols with a higher visibility in the alias.
- 
+- Add `--max-macro-iterations` to set macro iteration limit.
+
 ### Fixes
 - Bug in `io::write_using_write_byte`.
 - Bitstruct value cannot be used to index a const array in compile time. #2512